+++ Datalib/src/main/java/com/ing/datalib/or/mobile/MobileORObject.java
@@ -472,8 +472,12 @@ public class MobileORObject extends UndoRedoModel implements ORObjectInf {
     @JsonIgnore
     @Override
     public Boolean rename(String newName) {
+        boolean renamedParent = true;
         if (getParent().getChildCount() == 1) {
-            getParent().rename(newName);
+            renamedParent = getParent().rename(newName);
+            if (!renamedParent) {
+                return false;
+            }
         }
         if (newName == null || newName.isBlank()) {
             return false;